requirements.txt
opencv-python
numpy
pydub
librosa
google-auth
google-auth-oauthlib
google-api-python-client
